added authorization to post put delete endpoints in tags

+++ projekt2/src/tags.ts
@@ -32,10 +32,11 @@ router.get('/tag/:id', (req: Request, res: Response) => {
 })
 
 router.post('/tag/', authMiddleware, (req: Request, res: Response) => {
-    if (! req.body) {
-        res.status(400).send({'err': 'no data provided to create a tag'})
-    }
-    else if (! req.body.name) {
+    //f (! req.body) {
+    //    res.status(400).send({'err': 'no data provided to create a tag'})
+    //} 
+    //this is never reached
+    if (! req.body.name) {
         res.status(400).send({'err': 'tag needs name'})
     } 
     else {
@@ -47,7 +48,7 @@ router.post('/tag/', authMiddleware, (req: Request, res: Response) => {
             res.status(400).send({'err': 'you need to be logged in to post notes'})
           }
         if (tag_id != undefined) {
-            res.status(200).send({'id' : tag_id})
+            res.status(200).send({'err': 'tag with this name already exists'});
         }
         else { 
             //only accepts unique tags from authenticated users
@@ -58,30 +59,58 @@ router.post('/tag/', authMiddleware, (req: Request, res: Response) => {
     }
 })
 
-router.put('/tag/:id', (req: Request, res: Response) => {
-    let tag: Tag = req.body;
+router.put('/tag/:id', authMiddleware, (req: Request, res: Response) => {
+    let logged_user: UserInfo = req.body.user;
+    delete req.body.user;
+    let new_tag: Tag = req.body;
     let id = +req.params.id;
     if (!tags.hasTag(id)) {
         res.status(404).send({'err': 'tag with this id not found'})
     }
-    let tag_id = tags.findTagId(tag.name);
+    let tag = tags.getTag(id);
+    if (tag != undefined){
+        let tag_id = tags.findTagId(new_tag.name);
     if (tag_id != undefined) {
         res.status(400).send({'error' : `'${tag.name}' already used by id:'${tag_id}'`})
     }
     else {
-        tag.id = id;
-        tags.updateTag(tag);
-        res.status(204).send({'id': id})
+        if(tag.owner_name == logged_user.name ||
+            logged_user.is_admin == true){
+            new_tag.id = id;
+            new_tag.owner_name = tag.owner_name;
+            tags.updateTag(new_tag);
+            res.status(204).send({'id': id})
+        }
+        else{
+            res.status(404).send({'err': 'you dont have permission to edit this note'})
+        }
+        
     }
+    }
+    else{
+        res.status(404).send({'err': 'tag with this id not found'})
+    }
+    
 })
 
-router.delete('/tag/:id', (req: Request, res: Response) => {
+router.delete('/tag/:id', authMiddleware, (req: Request, res: Response) => {
     let id = +req.params.id;
+    let logged_user: UserInfo = req.body.user;
     if (!tags.hasTag(id)) {
         res.status(404).send({'err': 'tag with this id not found'})
     }
-    tags.deleteTag(id);
-    res.status(204).send();
+    else{
+        let tag_to_delete = tags.getTag(id);
+        if(tag_to_delete && 
+            tag_to_delete.owner_name == logged_user.name ||
+            logged_user.is_admin == true) {
+                tags.deleteTag(id);
+                res.status(204).send();
+        }
+        else {
+            res.status(404).send({'err': 'you dont have permission to edit this note'});
+        }
+    }
 })
 
 export function process_tags(new_or_existing: Tag[]): Tag[] {
